# openapi-qr-review-v0.1.yaml
openapi: 3.0.3
info:
  title: Akrilik QR Review API
  version: 0.1.0
  description: 'Kontrak draft MVP v0.1 berdasarkan PRD 2 Oktober 2026. Bukan API yang sudah diimplementasikan. Endpoint
    /api/v1 dipakai browser Next.js atau backend pengganti di masa depan; /r/{token} tetap URL publik cetak. Server
    URL contoh harus diganti domain sendiri.


    Admin: Authorization Bearer access token Supabase Auth. Server verifikasi signature, expiry, issuer dan audience
    lalu periksa role admin dari sumber terpercaya; jangan percaya role user_metadata. Login/logout/refresh Supabase
    Auth di luar kontrak aplikasi ini. Jangan mengirim secret/service-role key dari browser.


    Owner: cookie opaque __Host-owner_session bound ke satu QR, ownership_id dan auth_generation. HttpOnly, Secure,
    SameSite=Strict, Path=/; expiry 30 menit. API owner pada origin yang sama dengan frontend. Unsafe cookie requests
    memerlukan Origin allowlisted dan X-CSRF-Token. Public POST login/aktivasi/claim memerlukan Origin allowlisted
    juga (untuk login CSRF); pembatasan origin tidak menggantikan verifikasi kode/PIN dan bukan bukti kepemilikan.
    Seluruh endpoint memaksa HTTPS.


    Semua body JSON menolak field tak dikenal dan null kecuali schema nullable; input tak ada dan null berbeda.
    Body maksimal 16 KiB (di luar layanan Auth). Timestamp RFC3339 UTC; id UUID; pin string empat digit termasuk
...


...


...
    menyimpan body sensitif; sumber IP untuk limiter hanya dari proxy terpercaya.


...


...


...
servers:
...
paths:
...
  /api/v1/admin/batches:
    post:
      operationId: createBatch
      summary: Buat batch QR unik
      description: Semua unit, hash kode dan snapshot terenkripsi rahasia 24 jam dibuat atomik sebelum 201. Tidak
        mengirim raw code dalam JSON batch. Jika gagal, tidak ada batch parsial. Label tidak harus unik. Status
        unit GENERATED dan QR UNACTIVATED.
      tags:
      - Batch
      security:
      - AdminBearer: []
      parameters:
      - $ref: '#/components/parameters/IdempotencyKey'
      responses:
        '201':
          description: Berhasil.
          headers:
            Cache-Control:
              description: 'Semua JSON, HTML dan unduhan: no-store. Semua resolver responses: no-store, max-age=0.
                CDN harus bypass cache.'
              schema:
                type: string
                example: no-store
            X-Request-Id:
              description: ID tracing nonrahasia; sama dengan request_id JSON.
              schema:
                type: string
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BatchResponse'
        '400':
          $ref: '#/components/responses/Error400'
        '401':
          $ref: '#/components/responses/Error401'
        '403':
          $ref: '#/components/responses/Error403'
        '409':
          $ref: '#/components/responses/Error409'
        '415':
          $ref: '#/components/responses/Error415'
        '422':
          $ref: '#/components/responses/Error422'
        '429':
          $ref: '#/components/responses/Error429'
        '500':
          $ref: '#/components/responses/Error500'
        '503':
          $ref: '#/components/responses/Error503'
        '413':
          $ref: '#/components/responses/Error413'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/BatchCreate'
            example:
              label: Produksi Oktober 2026
              quantity: 100
    get:
      operationId: listBatches
      summary: Daftar batch
      description: ''
      tags:
      - Batch
      security:
      - AdminBearer: []
      parameters:
      - $ref: '#/components/parameters/Limit'
      - $ref: '#/components/parameters/Cursor'
      responses:
        '200':
          description: Berhasil.
          headers:
            Cache-Control:
              description: 'Semua JSON, HTML dan unduhan: no-store. Semua resolver responses: no-store, max-age=0.
                CDN harus bypass cache.'
              schema:
                type: string
                example: no-store
            X-Request-Id:
              description: ID tracing nonrahasia; sama dengan request_id JSON.
              schema:
                type: string
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BatchListResponse'
        '400':
          $ref: '#/components/responses/Error400'
        '401':
          $ref: '#/components/responses/Error401'
        '403':
          $ref: '#/components/responses/Error403'
        '429':
          $ref: '#/components/responses/Error429'
        '500':
          $ref: '#/components/responses/Error500'
        '503':
          $ref: '#/components/responses/Error503'
  /api/v1/admin/batches/{batch_id}:
    get:
      operationId: getBatch
      summary: Detail batch
      description: ''
      tags:
      - Batch
      security:
      - AdminBearer: []
      parameters:
      - $ref: '#/components/parameters/BatchId'
      responses:
        '200':
          description: Berhasil.
          headers:
            Cache-Control:
              description: 'Semua JSON, HTML dan unduhan: no-store. Semua resolver responses: no-store, max-age=0.
                CDN harus bypass cache.'
              schema:
                type: string
                example: no-store
            X-Request-Id:
              description: ID tracing nonrahasia; sama dengan request_id JSON.
              schema:
                type: string
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BatchResponse'
        '401':
          $ref: '#/components/responses/Error401'
        '403':
          $ref: '#/components/responses/Error403'
        '404':
          $ref: '#/components/responses/Error404'
        '429':
          $ref: '#/components/responses/Error429'
        '500':
          $ref: '#/components/responses/Error500'
        '503':
          $ref: '#/components/responses/Error503'
  /api/v1/admin/batches/{batch_id}/exports:
    post:
      operationId: createExport
      summary: Antrekan ZIP produksi atau kode rahasia
      description: 'PUBLIC_QR: ZIP berisi manifest.csv dan QR PNG/SVG URL publik, tanpa kode rahasia. ACTIVATION_CODES:
        ZIP terpisah berisi activation-codes.csv (token,activation_code). Opsi gambar hanya berlaku PUBLIC_QR; bila
        disertakan pada ACTIVATION_CODES, 422. Kode rahasia hanya dapat diekspor sebelum snapshot 24 jam berakhir
        dan sebelum ada unit batch yang aktif, retired atau kodenya dirotasi. Snapshot yang tidak lagi valid: 409
        ACTIVATION_EXPORT_UNAVAILABLE. Pekerja harus memeriksa ulang sebelum membuat dan mengunduh arsip. Secret
        tidak melewati URL publik. Jangan memasukkan arsip rahasia ke vendor cetak QR.'
      tags:
      - Batch
      security:
      - AdminBearer: []
      parameters:
      - $ref: '#/components/parameters/BatchId'
      - $ref: '#/components/parameters/IdempotencyKey'
      responses:
        '202':
          description: Berhasil.
          headers:
            Cache-Control:
              description: 'Semua JSON, HTML dan unduhan: no-store. Semua resolver responses: no-store, max-age=0.
                CDN harus bypass cache.'
              schema:
                type: string
                example: no-store
            X-Request-Id:
              description: ID tracing nonrahasia; sama dengan request_id JSON.
              schema:
                type: string
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ExportJobResponse'
        '400':
          $ref: '#/components/responses/Error400'
        '401':
          $ref: '#/components/responses/Error401'
        '403':
          $ref: '#/components/responses/Error403'
        '404':
          $ref: '#/components/responses/Error404'
        '409':
          $ref: '#/components/responses/Error409'
        '410':
          $ref: '#/components/responses/Error410'
        '415':
          $ref: '#/components/responses/Error415'
        '422':
          $ref: '#/components/responses/Error422'
        '429':
          $ref: '#/components/responses/Error429'
        '500':
          $ref: '#/components/responses/Error500'
        '503':
          $ref: '#/components/responses/Error503'
        '413':
          $ref: '#/components/responses/Error413'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ExportCreate'
...
  /api/v1/admin/exports/{export_id}/download:
    get:
      operationId: downloadExport
      summary: Unduh ZIP melalui server
      description: 'READY saja; QUEUED/RUNNING/FAILED: 409 EXPORT_NOT_READY; expired: 410 EXPORT_EXPIRED. Semua
        arsip TTL maksimal 24 jam dari job dan arsip kode tidak melewati TTL snapshot batch. Download diproxy setelah
        otorisasi; tidak memberikan signed URL publik. Tidak mendukung range/resume pada MVP.'
      tags:
      - Batch
      security:
      - AdminBearer: []
      parameters:
      - $ref: '#/components/parameters/ExportId'
      responses:
        '200':
          description: Berhasil.
          headers:
            Cache-Control:
              description: 'Semua JSON, HTML dan unduhan: no-store. Semua resolver responses: no-store, max-age=0.
                CDN harus bypass cache.'
              schema:
                type: string
                example: no-store
            X-Request-Id:
              description: ID tracing nonrahasia; sama dengan request_id JSON.
              schema:
                type: string
            Content-Disposition:
              schema:
                type: string
                example: attachment; filename="qr-public.zip"
          content:
            application/zip:
              schema:
                type: string
                format: binary
        '401':
          $ref: '#/components/responses/Error401'
        '403':
          $ref: '#/components/responses/Error403'
        '404':
          $ref: '#/components/responses/Error404'
        '409':
          $ref: '#/components/responses/Error409'
        '410':
          $ref: '#/components/responses/Error410'
        '429':
          $ref: '#/components/responses/Error429'
        '500':
          $ref: '#/components/responses/Error500'
        '503':
          $ref: '#/components/responses/Error503'
...
  /api/v1/admin/qr-codes/{qr_id}/stock:
    patch:
      operationId: updateStock
      summary: Catat QC tersedia atau rusak
      description: 'GENERATED → AVAILABLE atau DAMAGED; AVAILABLE → DAMAGED. QR harus UNACTIVATED. DAMAGED terminal
        untuk stok, SOLD tidak dapat diubah di sini. Pada DAMAGED server juga mengubah QR ke RETIRED dan mencabut
        kode; semua dilakukan atomik. Tindakan ke status yang sama: 409 STOCK_TRANSITION_INVALID.'
      tags:
      - Inventory
      security:
      - AdminBearer: []
      parameters:
      - $ref: '#/components/parameters/QrId'
      - $ref: '#/components/parameters/IfMatch'
      responses:
        '200':
          description: Berhasil.
          headers:
            Cache-Control:
              description: 'Semua JSON, HTML dan unduhan: no-store. Semua resolver responses: no-store, max-age=0.
                CDN harus bypass cache.'
              schema:
                type: string
                example: no-store
            X-Request-Id:
              description: ID tracing nonrahasia; sama dengan request_id JSON.
              schema:
                type: string
            ETag:
              description: Versi QR resource; format strong ETag seperti "v7". Gunakan persis pada If-Match.
              schema:
                type: string
                example: '"v7"'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/QrAdminResponse'
        '400':
          $ref: '#/components/responses/Error400'
        '401':
          $ref: '#/components/responses/Error401'
        '403':
          $ref: '#/components/responses/Error403'
        '404':
          $ref: '#/components/responses/Error404'
        '409':
          $ref: '#/components/responses/Error409'
        '412':
          $ref: '#/components/responses/Error412'
        '415':
          $ref: '#/components/responses/Error415'
        '422':
          $ref: '#/components/responses/Error422'
        '428':
          $ref: '#/components/responses/Error428'
        '429':
          $ref: '#/components/responses/Error429'
        '500':
          $ref: '#/components/responses/Error500'
        '503':
          $ref: '#/components/responses/Error503'
        '413':
          $ref: '#/components/responses/Error413'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/StockUpdate'
  /api/v1/admin/qr-codes/{qr_id}/sales:
    post:
      operationId: recordSale
      summary: Catat unit terjual
      description: Hanya AVAILABLE + UNACTIVATED. Create sale dan ubah stok SOLD dalam transaksi. Satu sale per
        QR pada MVP. Reference boleh sama untuk beberapa unit satu transaksi. sold_at tidak boleh di masa depan.
        Retry idem replay dilakukan sebelum evaluasi If-Match. Tidak ada payment gateway.
      tags:
      - Inventory
      security:
      - AdminBearer: []
      parameters:
      - $ref: '#/components/parameters/QrId'
      - $ref: '#/components/parameters/IfMatch'
      - $ref: '#/components/parameters/IdempotencyKey'
      responses:
        '201':
          description: Berhasil.
          headers:
            Cache-Control:
              description: 'Semua JSON, HTML dan unduhan: no-store. Semua resolver responses: no-store, max-age=0.
                CDN harus bypass cache.'
              schema:
                type: string
                example: no-store
            X-Request-Id:
              description: ID tracing nonrahasia; sama dengan request_id JSON.
              schema:
                type: string
            ETag:
              description: Versi QR resource; format strong ETag seperti "v7". Gunakan persis pada If-Match.
              schema:
                type: string
                example: '"v7"'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SaleResponse'
        '400':
          $ref: '#/components/responses/Error400'
        '401':
          $ref: '#/components/responses/Error401'
        '403':
          $ref: '#/components/responses/Error403'
        '404':
          $ref: '#/components/responses/Error404'
        '409':
          $ref: '#/components/responses/Error409'
        '412':
          $ref: '#/components/responses/Error412'
        '415':
          $ref: '#/components/responses/Error415'
        '422':
          $ref: '#/components/responses/Error422'
        '428':
          $ref: '#/components/responses/Error428'
        '429':
          $ref: '#/components/responses/Error429'
        '500':
          $ref: '#/components/responses/Error500'
        '503':
          $ref: '#/components/responses/Error503'
        '413':
          $ref: '#/components/responses/Error413'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SaleCreate'
    get:
      operationId: getSale
      summary: Detail penjualan unit
      description: 404 SALE_NOT_FOUND jika QR belum terjual. Kontak dukungan tidak diekspos pada API publik atau
        owner.
      tags:
      - Inventory
      security:
      - AdminBearer: []
      parameters:
      - $ref: '#/components/parameters/QrId'
      responses:
        '200':
          description: Berhasil.
          headers:
            Cache-Control:
              description: 'Semua JSON, HTML dan unduhan: no-store. Semua resolver responses: no-store, max-age=0.
                CDN harus bypass cache.'
              schema:
                type: string
                example: no-store
            X-Request-Id:
              description: ID tracing nonrahasia; sama dengan request_id JSON.
              schema:
                type: string
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SaleResponse'
        '401':
          $ref: '#/components/responses/Error401'
        '403':
          $ref: '#/components/responses/Error403'
        '404':
          $ref: '#/components/responses/Error404'
        '429':
          $ref: '#/components/responses/Error429'
        '500':
          $ref: '#/components/responses/Error500'
        '503':
          $ref: '#/components/responses/Error503'
...
  /api/v1/public/qr/{token}/activate:
    post:
      operationId: activateQr
      summary: Aktifkan unit toko
      description: 'Memerlukan UNACTIVATED + SOLD dan kode benar. Kode salah: 401 INVALID_ACTIVATION_CODE. Stok
        belum SOLD: 409 QR_NOT_SOLD; sudah aktif: 409 QR_ALREADY_ACTIVE; retired: 410 QR_RETIRED. Hash PIN, konsumsi
        kode, simpan toko dan buat ownership_id pertama atomik. Tidak membuat cookie; pemilik login via /owner/sessions.
        Aktivasi paralel hanya satu sukses. Retry key+payload yang sama mereplay hasil committed bila ownership/state
        belum berubah; jika berubah 409 QR_STATE_CHANGED. Key baru setelah aktif bukan replay.'
      tags:
      - Public
      security: []
      parameters:
      - $ref: '#/components/parameters/Token'
      - $ref: '#/components/parameters/IdempotencyKey'
      - $ref: '#/components/parameters/Origin'
      responses:
        '201':
          description: Berhasil.
          headers:
            Cache-Control:
              description: 'Semua JSON, HTML dan unduhan: no-store. Semua resolver responses: no-store, max-age=0.
                CDN harus bypass cache.'
              schema:
                type: string
                example: no-store
            X-Request-Id:
              description: ID tracing nonrahasia; sama dengan request_id JSON.
              schema:
                type: string
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ActivationResultResponse'
        '400':
          $ref: '#/components/responses/Error400'
        '401':
          $ref: '#/components/responses/Error401'
        '403':
          $ref: '#/components/responses/Error403'
        '404':
          $ref: '#/components/responses/Error404'
        '409':
          $ref: '#/components/responses/Error409'
        '410':
          $ref: '#/components/responses/Error410'
        '415':
          $ref: '#/components/responses/Error415'
        '422':
          $ref: '#/components/responses/Error422'
        '429':
          $ref: '#/components/responses/Error429'
        '500':
          $ref: '#/components/responses/Error500'
        '503':
          $ref: '#/components/responses/Error503'
        '413':
          $ref: '#/components/responses/Error413'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ActivationRequest'
            example:
              activation_code: MZXW6YTBOI2DEZLG
              store_name: Kopi Bahagia
              review_url: https://g.page/r/EXAMPLECODE/review
              pin: '0427'
              pin_confirmation: '0427'
...
  /api/v1/owner/me/stats:
    get:
      operationId: ownerStats
      summary: Statistik periode pemilik saat ini
      description: Total durable visits dan last_visited_at pada ownership_id saat ini. Dapat dibaca saat suspend.
        Angka bisa tertunda maksimal 60 detik; as_of menjelaskan kesegaran, kegagalan pencatatan bisa undercount.
        Bukan jumlah review atau orang unik.
      tags:
      - Owner
      security:
      - OwnerCookie: []
      parameters: []
      responses:
        '200':
          description: Berhasil.
          headers:
            Cache-Control:
              description: 'Semua JSON, HTML dan unduhan: no-store. Semua resolver responses: no-store, max-age=0.
                CDN harus bypass cache.'
              schema:
                type: string
                example: no-store
            X-Request-Id:
              description: ID tracing nonrahasia; sama dengan request_id JSON.
              schema:
                type: string
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StatsResponse'
        '401':
          $ref: '#/components/responses/Error401'
        '403':
          $ref: '#/components/responses/Error403'
        '429':
          $ref: '#/components/responses/Error429'
        '500':
          $ref: '#/components/responses/Error500'
        '503':
          $ref: '#/components/responses/Error503'
...
components:
  securitySchemes:
    AdminBearer:
      type: http
      scheme: bearer
      bearerFormat: Supabase JWT
      description: User access token valid plus authorization ADMIN; bukan service-role key.
    OwnerCookie:
      type: apiKey
      in: cookie
      name: __Host-owner_session
      description: Opaque session bound satu QR; dibentuk melalui login PIN.
...
  schemas:
...
    BatchResponse:
      type: object
      additionalProperties: false
      properties:
        data:
          $ref: '#/components/schemas/Batch'
        request_id:
          type: string
      required:
      - data
      - request_id
...